fix npm audit issues
The `pkg` package (used to build stand-alone binaries; not shipped with
ShenaniBot installs) depends on an old version of ansi-regex which has a
moderate "Inefficient Regular Expression Complexity" issue, which could
not be immediately resolved.

A reasonable threat assessment suggests this is a non-issue, but it's
still a nuisance since it will appear in our audit reports.  Will be
looking for an update to pkg (which might resolve the issue by updating
its dependencies), and may explore other options if the issue persists
for too long.